Added Agent 2 Model Based Reflex Agent
This agent will know what is the immediate environment and will take the action accordingly

If the agent see's that there is a energy down in any adjacent cell, it will avoid going there
If agent's Current Live is 10(Maximum), then agent will avoid going to Power Up as it cannot hold more power
If a agent has Life Jacket in any adjacent cell, it will grab it.
If the agent see a pitfall, it will Jump over it.